+++ featureclassifier.h
@@ -17,11 +17,11 @@ public:
     void train(QString positiveTrainingSet, QString negativeTrainingSet);
     void train(std::vector<cv::Mat1b> &truePositive, std::vector<cv::Mat1b> &falsePositive);
     bool classify(const cv::Mat1b &window) override;
-    bool classify(const cv::Mat1b &window, double *_c1, double *_c2, double *_c3, double *_c4);
+    bool classify(const cv::Mat1b &window, float *_c1, float *_c2, float *_c3, float *_c4);
 
 private:
-    double _t1, _t2, _tLower[4], _tUpper[4];
-    void _setConstants(const cv::Mat1b &gray, double *_c1, double *_c2, double *_c3, double *_c4);
+    float _t1, _t2, _tLower[4], _tUpper[4];
+    void _setConstants(const cv::Mat1b &gray, float *_c1, float *_c2, float *_c3, float *_c4);
     std::vector<cv::Point> _positiveMFICoordinates, _negativeMFICoordinates;
 };
